cleaned up home page so that the game gets hit first
