+++ frontend/src/app/JoinGame.tsx
